+++ src/main/java/darkoverload/itzip/global/config/response/code/ResponseCode.java
@@ -1,0 +1,9 @@
+package darkoverload.itzip.global.config.response.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface ResponseCode {
+    String name();
+    HttpStatus getHttpStatus();
+    String getData();
+}
