+++ src/main/java/darkoverload/itzip/global/config/response/exception/RestApiException.java
@@ -1,11 +1,11 @@
 package darkoverload.itzip.global.config.response.exception;
 
-import darkoverload.itzip.global.config.response.exceptionCode.ExceptionCode;
+import darkoverload.itzip.global.config.response.code.ResponseCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
 public class RestApiException extends RuntimeException {
-    private final ExceptionCode exceptionCode;
+    private final ResponseCode exceptionCode;
 }
