--- src/main/java/darkoverload/itzip/global/config/response/exceptionCode/ExceptionCode.java
@@ -1,9 +1,0 @@
-package darkoverload.itzip.global.config.response.exceptionCode;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-    String name();
-    HttpStatus getHttpStatus();
-    String getMessage();
-}
